Modify registrarSale to not include input calls + Creation of class: SaleReq

+++ src/product/Product.java
@@ -44,7 +44,7 @@ public class Product {
 	}
 
 
-	public float getCode() {
+	public int getCode() {
 		return code;
 	}
